+++ common/src/main/java/com/baisylia/modestmagic/config/ModConfig.java
@@ -17,6 +17,7 @@ public class ModConfig {
 
     public boolean throwItemsOnPedestals = true;
     public boolean reducedRrvMotion = false;
+    public boolean showTabletTooltips = true;
 
     public static ModConfig get() {
         if (INSTANCE == null) {