+++ common/src/main/java/com/baisylia/modestmagic/recipe/custom/TabletSmithingRecipe.java
@@ -88,7 +88,7 @@ public record TabletSmithingRecipe(Ingredient template, Ingredient base,
             Holder<Enchantment> enchantHolder = opt.get();
             Enchantment enchantment = enchantHolder.value();
 
-            if (enchantment.canEnchant(itemstack) && areEnchantsCompatible(existing, enchantHolder)) {
+            if (Services.PLATFORM.isPrimaryEnchantItem(itemstack, enchantHolder) && areEnchantsCompatible(existing, enchantHolder)) {
                 int currentLevel = mutable.getLevel(enchantHolder);
                 int targetLevel = currentLevel + 1;
                 if (targetLevel <= enchantment.getMaxLevel()) {