+++ CHANGELOG.md
@@ -5,6 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.3.1] - 2026-07-16
+
+### Fixed
+
+- Adjusted default enchantment compatibility.
+- Fixed auto-generated tooltip not working for certain enchantments.
+
+## [1.3.0] - 2026-06-06
+
+### Added
+
+- Added custom tooltips for tablets, with Enchantment Descriptions/Item Descriptions integration.
+
 ## [1.2.2] - 2026-05-13
 
 ### Changed